Bootanimation: Fix the low memory device oom when run
Low memory device is easily to be oom when run boot animation. Here
we do optimize the boot animation GL Texture only one frame needed
when free memory is not enough for bootanimation to run smoothly.
That is free timely and re-init it again. When in boot animation,
the fps=15, so it is ok to do the delete and re-init.

If the free memory is enough, the original design for bootanimation
will be much faster for each frame. But if free memory is not enough,
we will free the memory to avoid animation being stuck. This is the
balance design.

Change-Id: I315154557b4e43f9898d78ad06236f1804f6224b